Make constraining rect relative to the scroll container for sticky item
In the case "flex-direction: column-reverse" or "writing-mode: vertical-rl",
the offset of scrollable overflow rect may be not (0,0),
top or left could be negative.
If the sticky container is not the same with scroll container,
then everything is relative to scroll container,
so sticky item works fine.
But if sticky container is the same with scroll container,
then containing block rect and sticky box rect could have
negative offset, and constraining rect still starts with
offset (0,0), so the final sticky offset is not correct.

So if sticky container is scroll container, we have to
make sure constraining rect have the same offset with
scrollable overflow rect.

Bug: 331753412
Change-Id: I877b73afc458cd0de0ecfb618e33bc95118380a1